modifica HashFunction universale, e OPH con rotazione e reindex

+++ AlgoInC++/codice/Prime.cpp
@@ -1,19 +1,19 @@
 #include "Prime.h"
 
-bool Prime::isPrime(int n) {
+bool Prime::isPrime(size_t n) {
     if (n <= 1) return false;
     if (n == 2) return true;
     if (n % 2 == 0) return false;
     
     // Controlliamo solo i numeri dispari fino alla radice quadrata di n
-    int sqrtN = sqrt(n);
-    for (int i = 3; i <= sqrtN; i += 2) {
+    size_t sqrtN = sqrt(n);
+    for (size_t i = 3; i <= sqrtN; i += 2) {
         if (n % i == 0) return false;
     }
     return true;
 }
 
-int Prime::nextPrime(int num) {
+size_t Prime::nextPrime(size_t num) {
     // Se il numero è pari, partiamo dal successivo dispari
     if (num % 2 == 0) num++;
     else num += 2;
